refactor: move isthmus helpers to reth-optimism-consensus
