+++ crates/optimism/consensus/src/validation/isthmus.rs
@@ -2,6 +2,15 @@
 
 use crate::OpConsensusError;
 use alloy_consensus::BlockHeader;
+use alloy_primitives::{address, Address, B256};
+use core::fmt::Debug;
+use reth_storage_api::{errors::ProviderResult, StorageRootProvider};
+use reth_trie_common::HashedStorage;
+use revm::database::BundleState;
+
+/// The L2 contract `L2ToL1MessagePasser`, stores commitments to withdrawal transactions.
+pub const ADDRESS_L2_TO_L1_MESSAGE_PASSER: Address =
+    address!("4200000000000000000000000000000000000016");
 
 /// Verifies that `withdrawals_root` (i.e. `l2tol1-msg-passer` storage root since Isthmus) field is
 /// set in block header.
@@ -13,26 +22,60 @@ pub fn ensure_withdrawals_storage_root_is_some<H: BlockHeader>(
     Ok(())
 }
 
+/// Computes the storage root of predeploy `L2ToL1MessagePasser.sol`.
+///
+/// Uses state updates from block execution. See also [`withdrawals_root_prehashed`].
+pub fn withdrawals_root<DB: StorageRootProvider>(
+    state_updates: &BundleState,
+    state: DB,
+) -> ProviderResult<B256> {
+    // if l2 withdrawals transactions were executed there will be storage updates for
+    // `L2ToL1MessagePasser.sol` predeploy
+    withdrawals_root_prehashed(
+        state_updates
+            .state()
+            .get(&ADDRESS_L2_TO_L1_MESSAGE_PASSER)
+            .map(|acc| {
+                HashedStorage::from_plain_storage(
+                    acc.status,
+                    acc.storage.iter().map(|(slot, value)| (slot, &value.present_value)),
+                )
+            })
+            .unwrap_or_default(),
+        state,
+    )
+}
+
+/// Computes the storage root of predeploy `L2ToL1MessagePasser.sol`.
+///
+/// Uses pre-hashed storage updates of `L2ToL1MessagePasser.sol` predeploy, resulting from
+/// execution of L2 withdrawals transactions. If none, takes empty [`HashedStorage::default`].
+pub fn withdrawals_root_prehashed<DB: StorageRootProvider>(
+    hashed_storage_updates: HashedStorage,
+    state: DB,
+) -> ProviderResult<B256> {
+    state.storage_root(ADDRESS_L2_TO_L1_MESSAGE_PASSER, hashed_storage_updates)
+}
+
 /// Verifies block header field `withdrawals_root` against storage root of
 /// `L2ToL1MessagePasser.sol` predeploy post block execution.
 ///
 /// Takes state updates resulting from execution of block.
 ///
 /// See <https://specs.optimism.io/protocol/isthmus/exec-engine.html#l2tol1messagepasser-storage-root-in-header>.
-#[cfg(feature = "std")]
 pub fn verify_withdrawals_root<DB, H>(
-    state_updates: &revm::database::BundleState,
+    state_updates: &BundleState,
     state: DB,
     header: H,
 ) -> Result<(), OpConsensusError>
 where
-    DB: reth_storage_api::StorageRootProvider,
-    H: BlockHeader + core::fmt::Debug,
+    DB: StorageRootProvider,
+    H: BlockHeader + Debug,
 {
     let header_storage_root =
         header.withdrawals_root().ok_or(OpConsensusError::L2WithdrawalsRootMissing)?;
 
-    let storage_root = reth_optimism_storage::predeploys::withdrawals_root(state_updates, state)
+    let storage_root = withdrawals_root(state_updates, state)
         .map_err(OpConsensusError::L2WithdrawalsRootCalculationFail)?;
 
     if header_storage_root != storage_root {
@@ -52,24 +95,20 @@ where
 /// execution of block, if any. Otherwise takes empty [`reth_trie_common::HashedStorage::default`].
 ///
 /// See <https://specs.optimism.io/protocol/isthmus/exec-engine.html#l2tol1messagepasser-storage-root-in-header>.
-#[cfg(feature = "std")]
 pub fn verify_withdrawals_root_prehashed<DB, H>(
-    hashed_storage_updates: reth_trie_common::HashedStorage,
+    hashed_storage_updates: HashedStorage,
     state: DB,
     header: H,
 ) -> Result<(), OpConsensusError>
 where
-    DB: reth_storage_api::StorageRootProvider,
+    DB: StorageRootProvider,
     H: BlockHeader + core::fmt::Debug,
 {
     let header_storage_root =
         header.withdrawals_root().ok_or(OpConsensusError::L2WithdrawalsRootMissing)?;
 
-    let storage_root = reth_optimism_storage::predeploys::withdrawals_root_prehashed(
-        hashed_storage_updates,
-        state,
-    )
-    .map_err(OpConsensusError::L2WithdrawalsRootCalculationFail)?;
+    let storage_root = withdrawals_root_prehashed(hashed_storage_updates, state)
+        .map_err(OpConsensusError::L2WithdrawalsRootCalculationFail)?;
 
     if header_storage_root != storage_root {
         return Err(OpConsensusError::L2WithdrawalsRootMismatch {
